moved code from application_controller to scraper_controller
