+++ Gemfile
@@ -6,6 +6,7 @@ source 'https://rubygems.org'
  # #1
  group :production do
    gem 'pg'
+   gem 'rails_12factor'
  end
  
  # #2
