+++ app/helpers/scraper_helper.rb
@@ -1,0 +1,2 @@
+module ScraperHelper
+end
